+++ evaluate.py
@@ -109,7 +109,7 @@ def main(unused_argv):
         predictions = predictions[common.OUTPUT_TYPE]
         predictions = tf.reshape(predictions, shape=[-1])
         labels = tf.reshape(samples[common.LABEL], shape=[-1])
-        weights = tf.to_float(tf.not_equal(labels, dataset.ignore_label))
+        weights = tf.cast(tf.not_equal(labels, dataset.ignore_label), tf.float32)
 
         # Set ignore_label regions to label 0, because metrics.mean_iou requires
         # range of labels = [0, dataset.num_classes). Note the ignore_label regions
